% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{barley}
\alias{barley}
\title{Wedderburn Barley data}
\format{
Data frame
}
\usage{
data(barley)
}
\description{
This table gives data from Wedderburn's 1974 paper (Quasi-likelihood functions,
generalized linear models, and the Gauss-Newton method). It is the
the percentage leaf area of barley infected by the fungus R. Secalis (leaf blotch) for 10 different
varieties grown at 9 different sites in a variety trial in 1965. This data
was found in an unblished thesis of J. F. Jenkin.
}
\references{
Wedderburn, Quasi-likelihood functions, generalized linear models, and the Gauss-Newton method, Biometrika, 1974.
}
