% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.conf.int.r
\name{binom.conf.int}
\alias{binom.conf.int}
\title{Title}
\usage{
binom.conf.int(
  x,
  n,
  method = c("wald", "arcsine", "wilson", "clopper-pearson"),
  conf.level = 0.95,
  correct = FALSE
)
}
\arguments{
\item{correct}{}
}
\value{

}
\description{
Title
}
\details{
The Wilson and Clopper-Pearson procedures are obtained using \code{prop.test}
and \code{binom.test} respectively.
}
