% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.prediction.r
\name{binom.prediction}
\alias{binom.prediction}
\title{Binomial prediction interval}
\usage{
binom.prediction(n, p, level = 0.95)
}
\arguments{
\item{n}{number of Bernoulli trials}

\item{p}{probability of success}

\item{level}{level of confidence}
}
\description{
Binomial prediction interval
}
\examples{
p <- seq(0,1,length = 1001)
IP <- binom.prediction(19, p)
plot(p, 19*p, type = "n", ylab = "x", main = "prediction intervals")
polygon( c(p, rev(p)), c(IP[1,], rev(IP[2,])), col = "red")
}
