% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{EUR2170}
\alias{EUR2170}
\title{European individuals biometrical data}
\format{
Data frame
}
\usage{
data(EUR2170)
}
\description{
This is a data set of 2170 european individuals.
The variables are \code{birth} (the date of birth), \code{country}, \code{sex},
\code{height}, \code{weight}, and \code{age}.
}
