% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.anova.r
\name{fisher.anova}
\alias{fisher.anova}
\title{One-Way Fisher Anova}
\usage{
fisher.anova(formula, data = environment(formula))
}
\arguments{
\item{formula}{an object of class formula.}

\item{data}{an optional data frame containing the variables in the model.}
}
\value{
\code{fisher.anova} returns an object of class \code{"fisher.anova"}, which
is pretty-printed in a traditional way.
}
\description{
One-Way Fisher Anova
}
\details{
This functions performs a one-way Fisher anova.
}
\examples{
fisher.anova(Time ~ Groupe, Clearance)
# Comparison with `aov` R function
summary(aov(Time ~ Groupe, Clearance))
}
