% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{FR1643}
\alias{FR1643}
\title{French individuals biometrical data}
\format{
Data frame
}
\usage{
data(FR1643)
}
\description{
This is a data set of 1643 French individuals. Data are self-reported.
The variables are \code{sex}, \code{age}, \code{prof} (the profession),
\code{dept} (département in which the person is living), \code{dip} (last
diploma obtained), \code{children} (the number of children), \code{height}
(in cm), and \code{weight} (in kg).
}
