% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{Galton}
\alias{Galton}
\title{Galton Height Data}
\format{Data frame}
\usage{
data(Galton)
}
\description{
This data set is reconstitued from Galton data tables on heights in families.
The variables are
\itemize{
\item Family: The family that the child belongs to, labeled from 1 to 204 and 136A
\item Father: The father's height, in inches
\item Mother: The mother's height, in inches
\item Gender: The gender of the child, male (M) or female (F)
\item Height: The height of the child, in inches
\item Kids: The number of kids in the family of the child
}
}
\references{
\url{http://www.randomservices.org/random/data/Galton.html} and
\url{http://www.medicine.mcgill.ca/epidemiology/hanley/galton/}
}
\keyword{Galton}
