% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{GLWD}
\alias{GLWD}
\title{Global Lakes and Wetlands}
\format{
Data frame
}
\usage{
data(GLWF)
}
\description{
This data set gives the size and the coordinates of 248613 lakes and bodies of water,
combining the GLWD-1 and GLWD-2 datasets.
The variables are \code{id} (a numerical id), \code{type}, \code{name},
\code{area.km2}, \code{perim.km}, \code{long} and \code{lat}.
}
\references{
\url{https://www.worldwildlife.org/pages/global-lakes-and-wetlands-database}
}
