% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{Halley}
\alias{Halley}
\title{Halley mortality table}
\format{
Data frame
}
\usage{
data(Halley)
}
\description{
This historical table was established by Halley in 1693, using data
transmitted by Caspar Neumann. It gives the number of individuals
reaching a given age, on a total of 1328 births, in the city of Breslau
(now Wrocław, in Poland).
The variables are \code{age}, going from 1 to 84, and \code{number},
the number of individuals.
This table is in fact a population pyramid. Assuming population
stationarity, mortality rates can be deduced.
For ages above 107, Halley states only that there are a total of
107 individuals. A plausible completion of the table is given
in the examples section.
}
\examples{
# survival function
S <- c(1, Halley$number/1238)
plot(0:84, S, type = "l", xlab = "t", ylab = "S(t)")
# life expectancy
sum(S)
# completing the table above age 84
P <- c(Halley$number,18,16,14,13,11,9,8,6,5,3,2,1,1)
S <- c(1, P/1238)
sum(S)
}
\references{
Nicolas Bacaër, Histoires de mathématiques et de population, Cassini, 2008.
La table de mortalité d'E. Halley, présentée par J. Dupaquier,
Annales de démographie historique, 1976.
}
