% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{Odier}
\alias{Odier}
\title{Odier Mortality table}
\format{
Data frame
}
\usage{
data(Odier)
}
\description{
This table gives the number of deaths in Genève, by age and sex,
for the years 1778 and 1779.
}
\references{
Louis Odier, Extrait mortuaire de Genève pour les années 1778 et 1779,
Mémoires de la société établie à Genève pour l'encouragement des arts et de l'agriculture, 1780.
\url{https://books.google.fr/books?id=cBUGAAAAQAAJ&pg=PA144}
}
