% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bins.r
\name{plotbins}
\alias{plotbins}
\title{Plotting a heatmap for 2D data}
\usage{
plotbins(
  x,
  y,
  n.x = 11,
  n.y = 11,
  xlim = range(x),
  ylim = range(y),
  freq = TRUE,
  text.cex = 1,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{n.x}{number of bins for x}

\item{n.y}{number of bins for y}

\item{xlim}{limits for x}

\item{ylim}{limits for y}

\item{freq}{If \code{TRUE}, the number of data points in each bin is printed (if non zero)}

\item{text.cex}{\code{cex} parameter for printing the frequencies}

\item{...}{extra parameters for \code{image}}
}
\description{
Plotting a heatmap for 2D data
}
\details{
plots a heatmap of discretized data
}
\examples{
x <- rnorm(500)
y <- x + rnorm(500)
plotbins(x, y)

}
