% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fisher.anova.r
\name{print.fisher.anova}
\alias{print.fisher.anova}
\title{Pretty Print Fisher Anova}
\usage{
\method{print}{fisher.anova}(x)
}
\arguments{
\item{x}{an object of class \code{"fisher.anova"}}
}
\value{
Returns a \code{NULL} object.
}
\description{
Pretty Print Fisher Anova
}
