% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.age.r
\name{r.age}
\alias{r.age}
\title{Randow draws of ages}
\usage{
r.age(n, S)
}
\arguments{
\item{n}{number of draws}

\item{S}{numeric vector giving the survival function}
}
\value{
A vector of \code{n} random draws.
}
\description{
Randow draws of ages
}
\details{
The vector \code{S} gives the survival function.
Specifically, \code{ S[i] = P(T > i-1)} ; \code{S}
must be decreasing. If \code{S[1] = 1}, \eqn{P(T > 0) = 1} and
all draws are positive.
}
\examples{
# A survival function obtained from Halley's table
P <- c(Halley$number,18,16,14,13,11,9,8,6,5,3,2,1,1)
S <- c(1, P/1238)
A <- r.age(1e4, S)
mean(A)
}
