% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{Ribe}
\alias{Ribe}
\title{Ribe Cemetry Anthropological Data}
\format{
Data frame
}
\usage{
data(Ribe)
}
\description{
This table gives data from the Ribe Cemetry in Denmark, including the sex of individuals,
an estimation of their age, both femurs length, their height, and an indicator variable
of hypoplasia, estimated from theeth enamel possible.
}
\references{
Mette Wod, Height Estimation from Skeletal Remains, 2008. 
\url{http://www.adbou.dk/fileadmin/adbou/projektopgaver/ADBOU_linear_regression_Mette_Wodx.pdf}
\url{http://mirror.aidoh.dk/mettewod/wod.zip}
}
