% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.r
\name{ROC}
\alias{ROC}
\title{Plots A ROC Curve and computes AUC}
\usage{
ROC(formula, data, plot = TRUE, add = FALSE, ...)
}
\arguments{
\item{formula}{an object of class formula.}

\item{data}{an optional data frame containing the variables in the model.}

\item{plot}{if \code{TRUE}, the function plots the ROC curve}

\item{add}{if \code{TRUE}, the ROC curve is added to an existing plot}

\item{...}{extra parameters for the plot function}
}
\value{
an invisible list with components \code{x} (the specificity),
\code{y} (the sensitivity), and \code{AUC} (the AUC).
}
\description{
Plots A ROC Curve and computes AUC
}
\examples{
ROC(T21 ~ hCG, data = T21)
ROC(T21 ~ I(-AFP), T21, add = TRUE, lty = 2)
ROC(T21 ~ I(hCG/100 - 6.8*AFP + 51*age.gest + 33*age), data = T21, add = TRUE, col = "red")
abline(0,1,lty=3)
legend("bottomright", lty = c(1,2,1),
       col = c("black", "black", "red"), legend = c("hCG", "-AFP", "all vars"))
}
