% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstpet.r
\name{rstpet}
\alias{rstpet}
\title{Random St. Petersburg Paradox lottery draws}
\usage{
rstpet(size, log2 = FALSE)
}
\arguments{
\item{size}{the number of draws}

\item{log2}{Logical. If \code{TRUE}, returns log2 of results}
}
\value{
a vector of length \code{size}.
}
\description{
Random St. Petersburg Paradox lottery draws
}
\details{
In this lottery, the gain is 1 with a probability 1/2,
2 with probability 1/4, 4 with probability 1/8, and so on.
}
