% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.r
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, sample = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector.}

\item{sample}{logical. Compute sample skewness or apply correction?}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The value of the skewness. If \code{sample = FALSE} a correction for sample size is applied.
}
\description{
Skewness
}
\examples{
# skewness of the height in EUR2170 data
skewness(EUR2170$height, sample = TRUE)
# adults only
skewness(EUR2170$height[EUR2170$age > 20], sample = TRUE)
}
