% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{T21}
\alias{T21}
\title{Trisomy 21 data}
\format{
Data frame
}
\usage{
data(T21)
}
\description{
This data set is a case/control data set for T21 status.
#' The variables are
\itemize{
\item age Mother's age
\item age.gest Gestationnal age (in weeks)
\item AFP Alphafoetoprotein
\item hCG human chorionic gonadotropin
\item T21 T21 status
}
}
\references{
Françoise Jauzein, \url{http://acces.ens-lyon.fr/}
}
\keyword{T21}
