% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.bins.r
\name{tablebins}
\alias{tablebins}
\title{Contingency table after binning data}
\usage{
tablebins(x, y, n.x = 11, n.y = 11)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{n.x}{number of bins for x}

\item{n.y}{number of bins for y}
}
\value{
a table object
}
\description{
Contingency table after binning data
}
