% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.r
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(x, sample = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector.}

\item{sample}{logical. Compute sample variance or apply correction?}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The variance, with ou whithout sample correction.
}
\description{
Variance
}
\examples{
# variance of the height in EUR2170 data
variance(EUR2170$height, sample = TRUE)
# adults only
he <- subset(EUR2170, age > 20, "height")
variance(he, sample = TRUE)
# males and females separately
heM <- subset(EUR2170, age > 20 & sex == "M", "height")
heF <- subset(EUR2170, age > 20 & sex == "F", "height")
variance(heM, sample = TRUE)
variance(heF, sample = TRUE)
}
