% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.conf.int.r
\name{binom.conf.int}
\alias{binom.conf.int}
\title{Confidence interval for the binomial parameter p}
\usage{
binom.conf.int(
  x,
  n,
  proc = c("wald", "arcsine", "wilson", "clopper-pearson"),
  conf.level = 0.95,
  correct = FALSE
)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of Bernoulli trials}

\item{proc}{which confidence interval procedure to use}

\item{conf.level}{confidence level for the returned confidence interval}

\item{correct}{a logical indicating whether continuity correction should be applied}
}
\value{
a numeric vector of length 2, giving the bounds of the confidence interval
}
\description{
Confidence interval for the binomial parameter p
}
\details{
The Wald procedure is the textbook procedure for confidence intervals.
Arcsine procedure is obtained through the variance stabilizing transformation \eqn{arcsine(sqrt(p))}.
Wilson and Clopper-Pearson procedures are obtained through the functions \code{prop.test}
and \code{binom.test} respectively.
}
\examples{
binom.conf.int(4, 10, "wald")
binom.conf.int(4, 10, "wilson")
binom.conf.int(4, 10, "arcsine")
binom.conf.int(4, 10, "clopper")
}
