% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.conf.int.r
\name{binom.conf.int}
\alias{binom.conf.int}
\title{Confidence interval for the binomial parameter p}
\usage{
binom.conf.int(
  x,
  n,
  proc = c("wald", "arcsine", "wilson", "clopper-pearson"),
  conf.level = 0.95,
  correct = FALSE
)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of Bernoulli trials}

\item{proc}{which confidence interval procedure to use}

\item{conf.level}{confidence level for the returned confidence interval}

\item{correct}{a logical indicating whether continuity correction should be applied}
}
\description{
Confidence interval for the binomial parameter p
}
\details{
The Wilson and Clopper-Pearson procedures are obtained using \code{prop.test}
and \code{binom.test} respectively.
}
