% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.r
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, excess.kurtosis = FALSE, sample = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector.}

\item{excess.kurtosis}{logical. Compute excess kurtosis instead of kurtosis?}

\item{sample}{logical. Compute sample kurtosis or apply correction?}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The value of the (excess) kurtosis. If \code{sample = FALSE} a correction for sample size is applied.
}
\description{
Kurtosis
}
\examples{
# kurtosis of the height in EUR2170 data
kurtosis(EUR2170$height, sample = TRUE)
# adults only
he <- subset(EUR2170, age > 20, "height")
kurtosis(he, sample = TRUE)
# males and females separately
heM <- subset(EUR2170, age > 20 & sex == "M", "height")
heF <- subset(EUR2170, age > 20 & sex == "F", "height")
kurtosis(heM, sample = TRUE)
kurtosis(heF, sample = TRUE)
}
