% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.r
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, sample = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector.}

\item{sample}{logical. Compute sample skewness or apply correction?}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The value of the skewness. If \verb!sample = FALSE! a correction for sample size is applied.
}
\description{
Skewness
}
